Launch SuperSU from Settings
If SuperSU is installed, show an entry in Settings below
"Developer options", which launches SuperSU when tapped.
At boot time SuperSU will detect if com.android.settings
contains the resources from this patch, and if so it
will hide itself from the launcher. This was implemented
in v1.85, so we only show the entry if the installed
SuperSU version is at least that.

Also hide Superuser if SuperSU is present

Signed-off-by: Dan Trifan <jimsth@gmail.com>

Change-Id: I762d4b37ab5479cc392e0bd05d5719907d8a53c0

Conflicts:
	res/values/strings.xml
	src/com/android/settings/Settings.java